+++ source/common/tables/dbo.appInternalTasks.sql
@@ -21,6 +21,7 @@ CREATE TABLE [dbo].[appInternalTasks]
 	[descriptor]			[varchar](256)	NOT NULL,
 	[task_name]				[varchar](256)	NOT NULL,
 	[flg_actions]			[smallint]		NULL,
+	[is_resource_intensive]	[bit]			NOT NULL CONSTRAINT [DF_appInternalTasks_flg_resource_intensive] DEFAULT (0),
 	CONSTRAINT [PK_appInternalTasks] PRIMARY KEY  CLUSTERED 
 	(
 		[id]